% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-variable-selection.R
\name{compute_all_means}
\alias{compute_all_means}
\title{Compute all conditional means for Y given X_-j}
\usage{
compute_all_means(
  fitted_model,
  x,
  conditional_prob,
  support_x,
  lambda = "min",
  post_lasso = FALSE
)
}
\arguments{
\item{fitted_model}{Fitted regression model}

\item{x}{Query point}

\item{conditional_prob}{Conditional probability matrix (p-by-M)}

\item{support_x}{Support set of X}

\item{lambda}{Lambda regularization choice in cv.glmnet}

\item{post_lasso}{A logical value. If TRUE, post_lasso should be used and the
\code{fitted_model} should be a low-dimensional GLM}
}
\value{
Conditional probability vector
}
\description{
Compute all conditional means for Y given X_-j
}
