% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-variable-selection.R
\name{compute_all_means_efficient}
\alias{compute_all_means_efficient}
\title{Compute the conditional mean using power trick (efficient version)}
\usage{
compute_all_means_efficient(
  fitted_model,
  X,
  conditional_prob_mat,
  support_x,
  lambda = "min"
)
}
\arguments{
\item{fitted_model}{Fitted regression model}

\item{X}{data matrix of dim n-by-p}

\item{conditional_prob_mat}{Conditional probability matrix of dim n-by-(p*M) and M is the size of support_x}

\item{support_x}{Support set of X}

\item{lambda}{Lambda regularization choice in cv.glmnet}
}
\value{
Matrix of dim n-by-p
}
\description{
Compute the conditional mean using power trick (efficient version)
}
