% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-HMM.R
\name{compute_conditional_prob}
\alias{compute_conditional_prob}
\title{Compute conditional probability for observed SNPs}
\usage{
compute_conditional_prob(x, pInit, pEmit, Q)
}
\arguments{
\item{x}{Observation from HMM}

\item{pInit}{Initial probability vector of length K}

\item{pEmit}{Emission array, of dimension p-by-M-K. p is number of SNPs, M is support size of observation and K is support size of latent variable from a Markov Chain}

\item{Q}{Transition array, same dimension with pEmit}
}
\value{
A vector of length of total number of SNP
}
\description{
Compute conditional probability for observed SNPs
}
