% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-HMM.R
\name{compute_forward_prob_Rcpp}
\alias{compute_forward_prob_Rcpp}
\title{Rcpp implementation of forward prob matrix computation}
\usage{
compute_forward_prob_Rcpp(x, pInit, pEmit, Q)
}
\arguments{
\item{x}{Observation from HMM}

\item{pInit}{Initial probability vector of length K}

\item{pEmit}{Emission array, of dimension p-by-M-K. p is number of SNPs, M is support size of observation and K is support size of latent variable from a Markov Chain}

\item{Q}{Transition array, same dimension with pEmit}
}
\value{
Matrix of dimension p-by-M
}
\description{
Rcpp implementation of forward prob matrix computation
}
