% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_helpers.R
\name{d1.wcgf}
\alias{d1.wcgf}
\title{\code{d1.wcgf} is a function computing the derivative of the cumulant generating
function (CGF) of distributions, multiplied by a weight function, from the GLM family}
\usage{
d1.wcgf(s, P, W, fam)
}
\arguments{
\item{s}{The point where the CGF will be computed.}

\item{P}{A vector containing the parameter values of the family of distributions.}

\item{W}{A vector containing the weights.}

\item{fam}{The GLM family which includes the distribution whose CGF is being
evaluated (values can be \code{gaussian}, \code{binomial}, \code{poisson}, etc).}
}
\value{
CGF of the weighted distribution evaluated at \code{s}.
}
\description{
\code{d1.wcgf} is a function computing the derivative of the cumulant generating
function (CGF) of distributions, multiplied by a weight function, from the GLM family
}
\examples{
d1.wcgf(s = 5, P = c(1,1,2), W = c(1,1,1), fam = 'poisson')

}
