% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_helpers.R
\name{dCRT_dist}
\alias{dCRT_dist}
\title{method helpers will go here (e.g. solving saddlepoint equations)}
\usage{
dCRT_dist(n, fitted.val, fam)
}
\arguments{
\item{n}{The point where the CGF will be computed.}

\item{fitted.val}{A vector containing the fitted parameter values by
fitting a GLM to X on Z.}

\item{fam}{The GLM family which includes the distribution whose CGF is being
evaluated (values can be \code{gaussian}, \code{binomial}, \code{poisson}, etc).}
}
\value{
Simulated data from an appropriate distribution.
}
\description{
\code{dCRT_dist} is a function that returns simulated data from an appropriate
distribution depending on a specified  GLM family
}
\examples{
dCRT_dist(n = 5, fitted.val = c(1,2,1,1,1), fam = 'poisson')

}
