% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{dCRT_f}
\alias{dCRT_f}
\title{\code{dCRT_f} is a simulation helper function carrying out the dCRT test from the spacrt package.}
\usage{
dCRT_f(
  data,
  X_on_Z_fam = "binomial",
  Y_on_Z_fam = "negative.binomial",
  B = 10000
)
}
\arguments{
\item{data}{A named list with fields \code{X} (an nx1 vector for the predictor
variable of interest), \code{Y} (an nx1 response vector), and \code{Z}
(an nxp matrix of covariates).}

\item{X_on_Z_fam}{The GLM family for the regression of X on Z
(values can be \code{gaussian}, \code{binomial}, \code{poisson}, \code{negative.binomial}, etc).}

\item{Y_on_Z_fam}{The GLM family for the regression of Y on Z
(values can be \code{gaussian}, \code{binomial}, \code{poisson}, \code{negative.binomial}, etc).}

\item{B}{Number of resamples}
}
\value{
A named list with fields \code{left_side_p_value}, \code{right_side_p_value}, and \code{both_side_p_value}.
}
\description{
\code{dCRT_f} is a simulation helper function carrying out the dCRT test from the spacrt package.
}
\examples{
n <- 100; p <- 5; gamma_1 = 1; beta_0 = -3; beta_1 = 1; theta = 1; gamma_0 = -2
B <- 1000
expit <- function(theta)(exp(theta)/(1+exp(theta)))
Z <- as.matrix(rnorm(n = n, mean = 0, sd = 1))
X <- rbinom(n = n, size = 1, prob = expit(gamma_0 + gamma_1*Z))
Y <- MASS::rnegbin(n = n, mu = exp(beta_0 + beta_1*Z), theta = theta)
data <- list(X = X, Y = Y, Z = Z)
X_on_Z_fam <- "binomial"
Y_on_Z_fam <- "poisson"
results <- dCRT_f(data, X_on_Z_fam, Y_on_Z_fam, B)
}
