% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{dCRT_internal}
\alias{dCRT_internal}
\title{The distilled conditional randomization test.}
\usage{
dCRT_internal(
  data,
  X_on_Z_fam,
  Y_on_Z_fam,
  fitting_X_on_Z = "glm",
  fitting_Y_on_Z = "glm",
  fit_vals_X_on_Z_own = NULL,
  fit_vals_Y_on_Z_own = NULL,
  B = 2000
)
}
\arguments{
\item{data}{A named list with fields \code{X} (an nx1 vector for the predictor
variable of interest), \code{Y} (an nx1 response vector), and \code{Z}
(an nxp matrix of covariates).}

\item{X_on_Z_fam}{The GLM family for the regression of X on Z
(values can be \code{gaussian}, \code{binomial}, \code{poisson}, etc).}

\item{Y_on_Z_fam}{The GLM family for the regression of Y on Z
(values can be \code{gaussian}, \code{binomial}, \code{poisson}, etc).}

\item{fitting_X_on_Z}{The fitting method for the regression X on Z.}

\item{fitting_Y_on_Z}{The fitting method for the regression Y on Z.}

\item{fit_vals_X_on_Z_own}{Vector of fitted values for X on Z in case the user's custom method.
Works only if fitting_X_on_Z = 'own'}

\item{fit_vals_Y_on_Z_own}{Vector of fitted values for Y on Z in case the user's custom method.
Works only if fitting_Y_on_Z = 'own'}

\item{B}{The number of resamples to draw.}
}
\value{
A named list with fields \code{test_stat}, \code{left_side_p_value},
\code{right_side_p_value}, and \code{both_side_p_value}.
}
\description{
\code{dCRT_internal} is a function carrying out the dCRT based on GLMs for \code{X|Z} and \code{Y|Z}.
}
\examples{
n <- 80; p <- 2; B <- 100
data <- list(X = rbinom(n = n, size = 1, prob = 0.2),
             Y = rpois(n = n, lambda = 1),
             Z = matrix(rnorm(n = n*p, mean = 0, sd = 1), nrow = n, ncol = p))
X_on_Z_fam <- "binomial"
Y_on_Z_fam <- "poisson"
dCRT_internal(data, X_on_Z_fam, Y_on_Z_fam,
              fitting_X_on_Z = 'rf',
              fitting_Y_on_Z = 'glm',
              B = 2000)

}
