% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-HMM.R
\name{fastPhase_new}
\alias{fastPhase_new}
\title{Compile fastPhase from R}
\usage{
fastPhase_new(
  fp_path,
  X_file,
  out_path = NULL,
  K = 12,
  numit = 25,
  phased = FALSE,
  seed = 1,
  num_random_start = 1
)
}
\arguments{
\item{fp_path}{Path to fastPhase software}

\item{X_file}{Path to inp file}

\item{out_path}{Path to output file (can be either relative or absolute)}

\item{K}{Number of latent states in Markov chain}

\item{numit}{Number of iterations in EM algorithm}

\item{phased}{Haplotyp (0, 1) or SNP (0,1,2)}

\item{seed}{Seed for random start in fastPhase}

\item{num_random_start}{Number of random initializations in EM algorithm}
}
\value{
Output NULL if there is no error from fastPhase
}
\description{
Compile fastPhase from R
}
