% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-HMM.R
\name{fit_HMM}
\alias{fit_HMM}
\title{Fit the HMM to get estimated parameters using fastPhase}
\usage{
fit_HMM(fp_path, out_path, hashing_id, X_file, K)
}
\arguments{
\item{fp_path}{File path to the fastPhase software}

\item{out_path}{File path to the folder where estimation results will be saved}

\item{hashing_id}{A hashing sequence with length 20}

\item{X_file}{File path to the folder where inp file input is saved}

\item{K}{Number of states for the hidden variable}
}
\value{
A list of estimated parameters
}
\description{
Fit the HMM to get estimated parameters using fastPhase
}
