% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_helpers.R
\name{fit_models}
\alias{fit_models}
\title{Fit models to data}
\usage{
fit_models(
  data,
  X_on_Z_fam,
  Y_on_Z_fam,
  fitting_X_on_Z = "glm",
  fitting_Y_on_Z = "glm",
  fit_vals_X_on_Z_own = NULL,
  fit_vals_Y_on_Z_own = NULL
)
}
\arguments{
\item{data}{A named list with fields \code{X} (an nx1 vector for the predictor
variable of interest), \code{Y} (an nx1 response vector), and \code{Z}
(an nxp matrix of covariates).}

\item{X_on_Z_fam}{The GLM family for the regression of X on Z
(values can be \code{gaussian}, \code{binomial}, \code{poisson}, etc).}

\item{Y_on_Z_fam}{The GLM family for the regression of Y on Z
(values can be \code{gaussian}, \code{binomial}, \code{poisson}, etc).}

\item{fitting_X_on_Z}{The fitting method for the regression X on Z.}

\item{fitting_Y_on_Z}{The fitting method for the regression Y on Z.}

\item{fit_vals_X_on_Z_own}{Fitted values of X_on_Z based on user's own method.}

\item{fit_vals_Y_on_Z_own}{Fitted values of Y_on_Z based on user's own method.}
}
\value{
Fitted values.
}
\description{
\code{fit_models} is a function carrying out the saddlepoint approximation to the
dCRT based on GLMs for \code{X|Z} and \code{Y|Z}.
}
\keyword{internal}
