% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{GCM}
\alias{GCM}
\title{The generalized covariance measure test of Shah and Peters.}
\usage{
GCM(
  data,
  X_on_Z_fam = NULL,
  Y_on_Z_fam = NULL,
  fit_glm_X = TRUE,
  fit_glm_Y = TRUE,
  aux_info_X_on_Z = NULL,
  aux_info_Y_on_Z = NULL
)
}
\arguments{
\item{data}{A named list with fields \code{X} (an nx1 vector for the predictor
variable of interest), \code{Y} (an nx1 response vector), and \code{Z}
(an nxp matrix of covariates).}

\item{X_on_Z_fam}{The GLM family for the regression of X on Z
(values can be \code{gaussian}, \code{binomial}, \code{poisson}, \code{negative.binomial}, etc).}

\item{Y_on_Z_fam}{The GLM family for the regression of Y on Z
(values can be \code{gaussian}, \code{binomial}, \code{poisson}, \code{negative.binomial}, etc).}

\item{fit_glm_X}{A logical variable indicating whether to use GLM for estimating E\link{X|Z}.
Default is TRUE. If set to FALSE, mean of X is used to estimate E\link{X|Z}.}

\item{fit_glm_Y}{A logical variable indicating whether to use GLM for estimating E\link{Y|Z}.
Default is TRUE. If set to FALSE, mean of Y is used to estimate E\link{Y|Z}.}

\item{aux_info_X_on_Z}{The auxiliary information that may be used for complex GLM regression
(For instance, when X_on_Z_fam = negative.binomial, the dispersion parameter should be provided).}

\item{aux_info_Y_on_Z}{The auxiliary information that may be used for complex GLM regression
(For instance, when Y_on_Z_fam = negative.binomial, the dispersion parameter should be provided).}
}
\value{
A named list with fields \code{test_stat} and \code{p_value}.
}
\description{
\code{GCM} is a function carrying out the GCM test based on GLMs for X|Z and Y|Z.
}
\examples{
n <- 20; p <- 2
data <- list(X = rbinom(n = n, size = 1, prob = 0.2),
             Y = rpois(n = n, lambda = 1),
             Z = matrix(rnorm(n = n*p, mean = 0, sd = 1), nrow = n, ncol = p))
X_on_Z_fam <- "binomial"
Y_on_Z_fam <- "poisson"
results <- GCM(data, X_on_Z_fam, Y_on_Z_fam)
results$test_stat
results$p_value
}
