% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-variable-selection.R
\name{GCM_HMM}
\alias{GCM_HMM}
\title{Compute GCM p-values}
\usage{
GCM_HMM(data)
}
\arguments{
\item{data}{A list consisting of data, the necessary HMM parameters, file paths and hashing_id}
}
\value{
A list with two resulting p-values obtained using different regularization parameters
}
\description{
Compute GCM p-values
}
