% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-HMM.R
\name{help_dgp}
\alias{help_dgp}
\title{Generate emission, transition and initial probabilites}
\usage{
help_dgp(
  K,
  M,
  p,
  gamma = 0.5,
  stay_prob,
  initial_prob = rep(1/K, K),
  alpha = NULL,
  beta = NULL,
  beta_prior = FALSE
)
}
\arguments{
\item{K}{Number of latent states}

\item{M}{Support size of haplotype}

\item{p}{Number of markers}

\item{gamma}{Emission probability for X being 0}

\item{stay_prob}{Stay probability for the Markov jump process}

\item{initial_prob}{Initial probability vector}

\item{alpha}{Shape1 parameter in beta prior}

\item{beta}{Shape2 parameter in beta prior}

\item{beta_prior}{A logical value. If TRUE, beta distribution will be used to generate emission probability}
}
\value{
List including final pEmit, Q and pInit
}
\description{
Generate emission, transition and initial probabilites
}
