% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-variable-selection.R
\name{knockoff_HMM}
\alias{knockoff_HMM}
\title{Compute knockoff rejection sets}
\usage{
knockoff_HMM(data)
}
\arguments{
\item{data}{A list consisting of data, the necessary HMM parameters, file paths and hashing_id}
}
\value{
A list of different rejection sets subject to different choice of regularization parameters
}
\description{
Compute knockoff rejection sets
}
