% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-HMM.R
\name{name_output}
\alias{name_output}
\title{rename the output from fastPhase}
\usage{
name_output(Q, pEmit, pInit)
}
\arguments{
\item{Q}{Transition probability array}

\item{pEmit}{Emission probability array}

\item{pInit}{Initial probability vector}
}
\value{
A list including the renamed inputs
}
\description{
rename the output from fastPhase
}
