% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_helpers.R
\name{nb_precomp}
\alias{nb_precomp}
\title{A function computing the dispersion parameter in negative binomial regression}
\usage{
nb_precomp(data)
}
\arguments{
\item{data}{A list containing the response Y and covariate Z}
}
\value{
a list containing the Poisson model fitted values and estimate for dispersion
}
\description{
A function computing the dispersion parameter in negative binomial regression
}
