% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-variable-selection.R
\name{post_lasso}
\alias{post_lasso}
\title{This is a post-lasso fitting function}
\usage{
post_lasso(
  X,
  Y,
  family = "binomial",
  lambda_param_vec = c("lambda.min", "lambda.1se")
)
}
\arguments{
\item{X}{A matrix including n rows and p columns}

\item{Y}{A vector of length n}

\item{family}{A GLM family}

\item{lambda_param_vec}{A vector including either "lambda.min" or "lambda.1se"}
}
\value{
A list of fitted models and active set selected from lasso algorithm
}
\description{
This is a post-lasso fitting function
}
