% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{score.test}
\alias{score.test}
\title{Score Test}
\usage{
score.test(data, X_on_Z_fam, Y_on_Z_fam)
}
\arguments{
\item{data}{A (non-empty) named list with fields \code{X} (an nx1 vector for the predictor
variable of interest), \code{Y} (an nx1 response vector), and \code{Z} (an nxp matrix
of covariates).}

\item{X_on_Z_fam}{The GLM family for the regression of X on Z (values can be \code{binomial},
\code{poisson}, etc).}

\item{Y_on_Z_fam}{The GLM family for the regression of Y on Z (values can be \code{binomial},
\code{poisson}, \code{negative.binomial}, etc).}
}
\value{
A named list with fields \code{test_stat}, \code{p.left} (Left-sided p-value),
\code{p.right} (Right-sided p-value), \code{p.both} (Two-sided p-value), and
\code{NB.disp.param}.
}
\description{
\code{score.test} is a function carrying out the score test based on GLMs for
\code{X|Z} and \code{Y|Z}.
}
\examples{
n <- 200; p <- 4
set.seed(1234)
X <- rbinom(n = n, size = 1, prob = 0.3)
Y <- rpois(n = n, lambda = 1)
Z <- matrix(rnorm(n = n*p, mean = 0, sd = 1), nrow = n, ncol = p)
data <- list(X = X, Y = Y, Z = Z)

res.score.1 <- score.test(data,
                          X_on_Z_fam = "binomial", Y_on_Z_fam = "negative.binomial")
res.score.1

}
