% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_helpers.R
\name{spa_cdf}
\alias{spa_cdf}
\title{\code{spa_cdf} SPA to CDF of T_n = S_n / sqrt(n)}
\usage{
spa_cdf(X, Y, X_on_Z_fit_vals, Y_on_Z_fit_vals, fam, R, max_expansions = 10)
}
\arguments{
\item{X}{The point where the CGF will be computed.}

\item{Y}{The point where the CGF will be computed.}

\item{X_on_Z_fit_vals}{X_on_Z_fit$fitted.values}

\item{Y_on_Z_fit_vals}{Y_on_Z_fit$fitted.values}

\item{fam}{The GLM family which includes the distribution whose CGF is being
evaluated (values can be \code{gaussian}, \code{binomial}, \code{poisson}, etc).}

\item{R}{stats::uniroot() search space endpoint}

\item{max_expansions}{Maximum number of times stats::uniroot() search space shuold be broadened}
}
\value{
Simulated data from an appropriate distribution.
}
\description{
\code{spa_cdf} SPA to CDF of T_n = S_n / sqrt(n)
}
\examples{
n <- 100; p <- 2; normalize <- FALSE; return_cdf <- FALSE
data <- list(X = rbinom(n = n, size = 1, prob = 0.2),
             Y = rpois(n = n, lambda = 1),
             Z = matrix(rnorm(n = n*p, mean = 0, sd = 1), nrow = n, ncol = p))
X <- data$X; Y <- data$Y; Z <- data$Z
X_on_Z_fit <- suppressWarnings(stats::glm(X ~ Z, family = "binomial"))
Y_on_Z_fit <- suppressWarnings(stats::glm(Y ~ Z, family = "poisson"))
spa_cdf(X = X, Y = Y, X_on_Z_fit_vals = X_on_Z_fit$fitted.values,
 Y_on_Z_fit_vals = Y_on_Z_fit$fitted.values, fam = "binomial", R = 1000)

}
