% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-variable-selection.R
\name{spaCRT_HMM}
\alias{spaCRT_HMM}
\title{Compute spaCRT p-values}
\usage{
spaCRT_HMM(data)
}
\arguments{
\item{data}{A list consisting of data, the necessary HMM parameters, file paths and hashing_id}
}
\value{
A list with two resulting p-values obtained using different regularization parameters
}
\description{
Compute spaCRT p-values
}
