% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-variable-selection.R
\name{compute_all_means}
\alias{compute_all_means}
\title{Compute all conditional means for Y|X_{-j}}
\usage{
compute_all_means(fitted_model, x, conditional_prob, support_x, lambda = "min")
}
\arguments{
\item{fitted_model}{Fitted regression model}

\item{x}{Query point}

\item{conditional_prob}{Conditional probability matrix (p-by-M)}

\item{support_x}{Support set of X}

\item{lambda}{Lambda regularization choice in cv.glmnet}
}
\value{
Conditional probability vector E(Y|X_{-j} = x_{-j})
}
\description{
Compute all conditional means for Y|X_{-j}
}
