% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{dCRT_internal}
\alias{dCRT_internal}
\title{The distilled conditional randomization test.}
\usage{
dCRT_internal(
  data,
  X_on_Z_fam,
  Y_on_Z_fam,
  fitting_X_on_Z = "glm",
  fitting_Y_on_Z = "glm",
  fit_vals_X_on_Z_own = NULL,
  fit_vals_Y_on_Z_own = NULL,
  B = 5000
)
}
\arguments{
\item{data}{A (non-empty) named list with fields \code{X} (an nx1 vector for the predictor
variable of interest), \code{Y} (an nx1 response vector), and \code{Z} (an nxp matrix
of covariates).}

\item{X_on_Z_fam}{The GLM family for the regression of X on Z (values can be \code{binomial},
\code{poisson}, etc).}

\item{Y_on_Z_fam}{The GLM family for the regression of Y on Z (values can be \code{binomial},
\code{poisson}, \code{negative.binomial}, etc).}

\item{fitting_X_on_Z}{The fitting method for the regression X on Z (values can be \code{glm} (default),
\code{random_forest}, or \code{NA}).}

\item{fitting_Y_on_Z}{The fitting method for the regression Y on Z (values can be \code{glm} (default),
\code{random_forest}, or \code{NA}).}

\item{fit_vals_X_on_Z_own}{Vector of fitted values for X on Z in case the user's custom method.}

\item{fit_vals_Y_on_Z_own}{Vector of fitted values for Y on Z in case the user's custom method.}

\item{B}{The number of resamples to draw (Default value is 2000).}
}
\value{
A named list with fields \code{test_stat}, \code{p.left} (Left-sided p-value),
\code{p.right} (Right-sided p-value), \code{p.both} (Two-sided p-value).
}
\description{
\code{dCRT_internal} is a function carrying out the dCRT based on GLMs for \code{X|Z} and \code{Y|Z}.
}
\examples{
n <- 200; p <- 4
set.seed(1234)
X <- rbinom(n = n, size = 1, prob = 0.3)
Y <- rpois(n = n, lambda = 1)
Z <- matrix(rnorm(n = n*p, mean = 0, sd = 1), nrow = n, ncol = p)
data <- list(X = X, Y = Y, Z = Z)

# fit both models via GLM
res.dCRT.1 <- dCRT_internal(data,
                            X_on_Z_fam = "binomial", Y_on_Z_fam = "poisson",
                            fitting_X_on_Z = 'glm', fitting_Y_on_Z = 'glm',
                            fit_vals_X_on_Z_own = NULL, fit_vals_Y_on_Z_own = NULL)
res.dCRT.1

# custom fit for Y|Z, explicit method/family only for X|Z
user_fit_Y <- glm(Y ~ Z,
                  family = poisson(),
                  data = data.frame(Y = Y, Z = Z))$fitted.values |> unname()

res.dCRT.2 <- dCRT_internal(data,
                            X_on_Z_fam = "binomial", Y_on_Z_fam = NULL,
                            fitting_X_on_Z = 'random_forest', fitting_Y_on_Z = NULL,
                            fit_vals_X_on_Z_own = NULL, fit_vals_Y_on_Z_own = user_fit_Y,
                            B = 10000)
res.dCRT.2

}
