% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_helpers.R
\name{fit_models}
\alias{fit_models}
\title{Fit models to data}
\usage{
fit_models(data, X_on_Z_fam, Y_on_Z_fam, fitting_method = "glm")
}
\arguments{
\item{data}{A named list with fields \code{X} (an nx1 vector for the predictor
variable of interest), \code{Y} (an nx1 response vector), and \code{Z}
(an nxp matrix of covariates).}

\item{X_on_Z_fam}{The GLM family for the regression of X on Z
(values can be \code{gaussian}, \code{binomial}, \code{poisson}, etc).}

\item{Y_on_Z_fam}{The GLM family for the regression of Y on Z
(values can be \code{gaussian}, \code{binomial}, \code{poisson}, etc).}

\item{fitting_method}{The fitting method for the regressions X on Z and Y on Z.}
}
\value{
Simulated data from an appropriate distribution.
}
\description{
\code{fit_models} is a function carrying out the saddlepoint approximation to the
dCRT based on GLMs for \code{X|Z} and \code{Y|Z}.
}
\examples{
n <- 100; p <- 5
data <- list(X = rbinom(n = n, size = 1, prob = 0.2),
             Y = rbinom(n = n, size = 1, prob = 0.7),
             Z = matrix(rnorm(n = n*p, mean = 0, sd = 1), nrow = n, ncol = p))
X_on_Z_fam <- "binomial"
Y_on_Z_fam <- "binomial"
fitting_method <- "rf"
fitted_vals <- fit_models(data, fitting_method, X_on_Z_fam, Y_on_Z_fam)
}
