% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{GCM}
\alias{GCM}
\title{The generalized covariance measure test of Shah and Peters.}
\usage{
GCM(data, X_on_Z_fam, Y_on_Z_fam, fitting_method = "glm")
}
\arguments{
\item{data}{A named list with fields \code{X} (an nx1 vector for the predictor
variable of interest), \code{Y} (an nx1 response vector), and \code{Z}
(an nxp matrix of covariates).}

\item{X_on_Z_fam}{The GLM family for the regression of X on Z
(values can be \code{gaussian}, \code{binomial}, \code{poisson}, \code{negative.binomial}, etc).}

\item{Y_on_Z_fam}{The GLM family for the regression of Y on Z
(values can be \code{gaussian}, \code{binomial}, \code{poisson}, \code{negative.binomial}, etc).}

\item{fitting_method}{The fitting method for the regressions X on Z and Y on Z.}
}
\value{
A named list with fields \code{test_stat} and \code{left_side_p_value},
\code{right_side_p_value}, \code{both_side_p_value}.
}
\description{
\code{GCM} is a function carrying out the GCM test based on GLMs for \code{X|Z} and \code{Y|Z}.
}
\examples{
n <- 20; p <- 2
data <- list(X = rbinom(n = n, size = 1, prob = 0.2),
             Y = rpois(n = n, lambda = 1),
             Z = matrix(rnorm(n = n*p, mean = 0, sd = 1), nrow = n, ncol = p))
X_on_Z_fam <- "binomial"
Y_on_Z_fam <- "poisson"
results <- GCM(data, X_on_Z_fam, Y_on_Z_fam, fitting_method = 'glm')

}
