% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-HMM.R
\name{help_dgp}
\alias{help_dgp}
\title{Generate emission, transition and initial probabilites}
\usage{
help_dgp(K, M, p, gamma, stay_prob, initial_prob = rep(1/K, K))
}
\arguments{
\item{K}{Number of latent states}

\item{M}{Support size of haplotype}

\item{p}{Number of markers}

\item{gamma}{Emission probability for X being 0}

\item{stay_prob}{Stay probability for the Markov jump process}

\item{initial_prob}{Initial probability vector}
}
\value{
List including final pEmit, Q and pInit
}
\description{
Generate emission, transition and initial probabilites
}
