% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{score.test}
\alias{score.test}
\title{Score Test}
\usage{
score.test(X, Y, Z, family)
}
\arguments{
\item{X}{Numeric vector of length \eqn{n}, representing the predictor variable.}

\item{Y}{Numeric vector of length \eqn{n}, representing the response variable.}

\item{Z}{Numeric matrix with \eqn{n} rows and \eqn{p} columns, representing covariates.}

\item{family}{Named list with elements \code{XZ} and \code{YZ} specifying the model family for \eqn{X \mid Z} and
\eqn{Y \mid Z} for each model. Each list element must be a string (e.g. \code{"binomial"},
\code{"poisson"}).}
}
\value{
A named list with fields \code{test_stat}, \code{p.left} (Left-sided p-value),
\code{p.right} (Right-sided p-value), \code{p.both} (Two-sided p-value), and
\code{NB.disp.param}.
}
\description{
\code{score.test} is a function carrying out the score test based on GLMs for
\code{X|Z} and \code{Y|Z}.
}
\examples{
n <- 200; p <- 4
set.seed(1234)
X <- rbinom(n = n, size = 1, prob = 0.3)
Y <- rpois(n = n, lambda = 1)
Z <- matrix(rnorm(n = n*p, mean = 0, sd = 1), nrow = n, ncol = p)

res.score.1 <- score.test(X = X, Y = Y, Z = Z,
                          family = list(XZ = "binomial", YZ = "negative.binomial"))
res.score.1

}
