% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{score.test}
\alias{score.test}
\title{Score Test}
\usage{
score.test(data, X_on_Z_fam = NULL, Y_on_Z_fam = NULL)
}
\arguments{
\item{data}{A named list with fields \code{X} (an nx1 vector for the predictor
variable of interest), \code{Y} (an nx1 response vector), and \code{Z}
(an nxp matrix of covariates).}

\item{X_on_Z_fam}{The GLM family for the regression of X on Z
(values can be \code{gaussian}, \code{binomial}, \code{poisson}, etc).}

\item{Y_on_Z_fam}{The GLM family for the regression of Y on Z
(values can be \code{gaussian}, \code{binomial}, \code{poisson}, etc).}
}
\value{
A named list with fields \code{test_stat} and \code{left_side_p_value},
\code{right_side_p_value} and \code{both_side_p_value}.
}
\description{
\code{score.test} is a function carrying out the saddlepoint approximation to the
dCRT based on GLMs for \code{X|Z} and \code{Y|Z}.
}
\examples{
n <- 50; p <- 2; normalize <- FALSE; return_cdf <- FALSE
data <- list(X = rbinom(n = n, size = 1, prob = 0.2),
             Y = rpois(n = n, lambda = 1),
             Z = matrix(rnorm(n = n*p, mean = 0, sd = 1), nrow = n, ncol = p))
X_on_Z_fam <- "binomial"
Y_on_Z_fam <- "negative.binomial"
results <- score.test(data, X_on_Z_fam, Y_on_Z_fam)
results$test_stat

}
