% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{spaCRT}
\alias{spaCRT}
\title{The saddlepoint approximation to the dCRT.}
\usage{
spaCRT(data, X_on_Z_fam, Y_on_Z_fam, normalize)
}
\arguments{
\item{data}{A named list with fields \code{X} (an nx1 vector for the predictor
variable of interest), \code{Y} (an nx1 response vector), and \code{Z}
(an nxp matrix of covariates).}

\item{X_on_Z_fam}{The GLM family for the regression of X on Z
(values can be \code{gaussian}, \code{binomial}, \code{poisson}, etc).}

\item{Y_on_Z_fam}{The GLM family for the regression of Y on Z
(values can be \code{gaussian}, \code{binomial}, \code{poisson}, etc).}

\item{normalize}{A logical variable indicating whether the spaCRT is based on
the normalized test statistic.}
}
\value{
A named list with fields \code{test_stat} and \code{p_value}.
}
\description{
\code{spaCRT} is a function carrying out the saddlepoint approximation to the
dCRT based on GLMs for X|Z and Y|Z.
}
