% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{spaCRT}
\alias{spaCRT}
\title{The saddlepoint approximation to the dCRT.}
\usage{
spaCRT(
  data,
  X_on_Z_fam = NULL,
  Y_on_Z_fam = NULL,
  normalize = FALSE,
  return_cdf = FALSE,
  fit_glm_X = TRUE,
  fit_glm_Y = TRUE,
  test_side = "right",
  aux_info_X_on_Z = NULL,
  aux_info_Y_on_Z = NULL
)
}
\arguments{
\item{data}{A named list with fields \code{X} (an nx1 vector for the predictor
variable of interest), \code{Y} (an nx1 response vector), and \code{Z}
(an nxp matrix of covariates).}

\item{X_on_Z_fam}{The GLM family for the regression of X on Z
(values can be \code{gaussian}, \code{binomial}, \code{poisson}, etc).}

\item{Y_on_Z_fam}{The GLM family for the regression of Y on Z
(values can be \code{gaussian}, \code{binomial}, \code{poisson}, etc).}

\item{normalize}{A logical variable indicating whether the spaCRT is based on
the normalized test statistic.}

\item{return_cdf}{A logical variable indicating whether to return the CDF}

\item{fit_glm_X}{A logical variable indicating whether to use GLM for estimating \code{E[X|Z]}.
Default is TRUE. If set to FALSE, mean of X is used to estimate \code{E[X|Z]}.}

\item{fit_glm_Y}{A logical variable indicating whether to use GLM for estimating \code{E[Y|Z]}.
Default is TRUE. If set to FALSE, mean of Y is used to estimate \code{E[Y|Z]}.}

\item{test_side}{A categorical variable indicating if the cnocerned test is a right-sided test,
or a left-sided test, or a both-sided test.}

\item{aux_info_X_on_Z}{The auxiliary information that may be used for complex GLM regression
(For instance, when X_on_Z_fam = negative.binomial, the dispersion parameter should be provided).}

\item{aux_info_Y_on_Z}{The auxiliary information that may be used for complex GLM regression
(For instance, when Y_on_Z_fam = negative.binomial, the dispersion parameter should be provided).}
}
\value{
A named list with fields \code{test_stat}, \code{left_side_p_value},
\code{right_side_p_value}, \code{both_side_p_value}, \code{cdf}, and
\code{gcm.default}.
Here, cdf is a function that takes in a value t and returns the
saddlepoint approximation to the CDF of the resampling distribution of the
test statistic evaluated at t. This function is returned only if
return_cdf == TRUE.
gcm.default returns TRUE if spacrt::GCM was employed due to the failure of spaCRT.
}
\description{
\code{spaCRT} is a function carrying out the saddlepoint approximation to the
dCRT based on GLMs for \code{X|Z} and \code{Y|Z}.
}
\examples{
n <- 20; p <- 2; normalize <- FALSE; return_cdf <- FALSE
data <- list(X = rbinom(n = n, size = 1, prob = 0.2),
             Y = rpois(n = n, lambda = 1),
             Z = matrix(rnorm(n = n*p, mean = 0, sd = 1), nrow = n, ncol = p))
X_on_Z_fam <- "binomial"
Y_on_Z_fam <- "poisson"
results <- spaCRT(data, X_on_Z_fam, Y_on_Z_fam, normalize)
results$test_stat
results$p_value

}
